fix: lsp--find-workspaces-for workspace/executeCommand respects the target command
Calling

    (lsp--find-workspaces-for '(:method "workspace/executeCommand"
                                :params (:command "arbitrary-command")))

would return *any* workspace that had the :executeCommandProvider capability
registerd.

This would result in a given command being forwarded for **all active
workspaces**. Some servers (such as ruff) would raise a KeyError for unknown
commands -- causing the minibuffer to, sometimes, show an error.

This change introduces a new :check-message callback that can be used in
lsp-method-requirements -- this callback receives the workspace and the
message to be sent.

Ths entry for workspace/executeCommand has been updated to check if the target
workspace can execute the required command.